#CKFK20230416 Modificacion a la comunicación

+++ app/src/main/java/com/dts/roadp/clsDataBuilder.java
@@ -128,8 +128,8 @@ public class clsDataBuilder {
 					if (ct==0) s=""+DT.getDouble(i);
 					if (ct==1) s="'"+DT.getString(i)+"'";
 
-					if (ct==2) s="'"+DU.univfechaext(DT.getInt(i))+"'";
-					if (ct==3) s="'"+DU.univfechaext(DT.getInt(i))+"'";
+					if (ct==2) s="'"+DU.univfechaext(DT.getLong(i))+"'";
+					if (ct==3) s="'"+DU.univfechaext(DT.getLong(i))+"'";
 					
 					if (i<cc-1) s=s+",";
 					si=si+s;
